moved writing functions to io_utils
