+++ src/main.rs
@@ -1,9 +1,9 @@
-use std::{
-    fs::OpenOptions,
-    io::{self, stdout, Write},
-};
+use std::io::{self, stdout, Write};
 
-use stats::{io_utils::read_tallies, mapping};
+use stats::{
+    io_utils::{read_tallies, save_popsync_results, save_tracking_results},
+    mapping,
+};
 
 fn main() {
     // Input handling
@@ -37,61 +37,4 @@ fn main() {
 
     save_tracking_results(&tracking_res);
     save_popsync_results(&popsync_res);
-}
-
-pub fn save_tracking_results(tracking_res: &[f64]) {
-    // The table is something like this
-    //
-    //               | Absorb | Scatter | Fission | Collision | Census | NumSeg
-    // CycleTracking | ...
-    //
-    let mut file = OpenOptions::new()
-        .write(true)
-        .create(true)
-        .truncate(true)
-        .open("tracking.dat")
-        .unwrap();
-    writeln!(file, ",Absorb,Scatter,Fission,Collision,Census,NumSeg").unwrap();
-    // write correlation coeffs
-    writeln!(
-        file,
-        "CycleTracking, {:.5}, {:.5}, {:.5}, {:.5}, {:.5}, {:.5}",
-        tracking_res[0],
-        tracking_res[1],
-        tracking_res[2],
-        tracking_res[3],
-        tracking_res[4],
-        tracking_res[5],
-    )
-    .unwrap();
-    // padding values for it to be considered a matrix
-    writeln!(file, "Dummy, 0, 0, 0, 0, 0, 0").unwrap();
-}
-
-pub fn save_popsync_results(popsync_res: &[f64]) {
-    // The table is something like this
-    //
-    //                   | Source | Rr | Split
-    // PopulationControl | ...
-    // CycleSync         | ...
-    //
-    let mut file = OpenOptions::new()
-        .write(true)
-        .create(true)
-        .truncate(true)
-        .open("popsync.dat")
-        .unwrap();
-    writeln!(file, ",Rr,Split").unwrap();
-    writeln!(
-        file,
-        "CycleSync, {:.5}, {:.5}",
-        popsync_res[1], popsync_res[2]
-    )
-    .unwrap();
-    writeln!(
-        file,
-        "PopulationControl, {:.5}, {:.5}",
-        popsync_res[4], popsync_res[5]
-    )
-    .unwrap();
 }
